move auth pages into shared subdirectory
